Added a route for updating the user's profile
Added a route for a patch request to allow the user to update its profile.
The route takes the user's token in order to get her id and find her and then update with the request body.
The authentication and verification of token is added as a middleware to the route.

Ref: Req #002